+++ lib/src/features/restrict_apps/method_channel/restrictions_method_channel.dart
@@ -5,7 +5,6 @@ import 'package:pauza_screen_time/src/core/app_identifier.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/app_restriction_platform.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/method_channel/channel_name.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/method_channel/method_names.dart';
-import 'package:pauza_screen_time/src/features/restrict_apps/model/restriction_schedule_config.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/model/restriction_scheduled_mode.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/model/restriction_scheduled_modes_config.dart';
 import 'package:pauza_screen_time/src/features/restrict_apps/model/restriction_session.dart';
@@ -128,32 +127,6 @@ class RestrictionsMethodChannel extends AppRestrictionPlatform {
     return channel.invokeMethod<void>(
       RestrictionsMethodNames.endRestrictionSession,
     );
-  }
-
-  @override
-  Future<void> setRestrictionScheduleConfig(RestrictionScheduleConfig config) {
-    return channel.invokeMethod<void>(
-      RestrictionsMethodNames.setRestrictionScheduleConfig,
-      config.toMap(),
-    );
-  }
-
-  @override
-  Future<RestrictionScheduleConfig> getRestrictionScheduleConfig() async {
-    final result = await channel.invokeMethod<Map<dynamic, dynamic>>(
-      RestrictionsMethodNames.getRestrictionScheduleConfig,
-    );
-    if (result == null) {
-      return const RestrictionScheduleConfig(enabled: false, schedules: []);
-    }
-
-    try {
-      return RestrictionScheduleConfig.fromMap(
-        Map<String, dynamic>.from(result),
-      );
-    } catch (_) {
-      return const RestrictionScheduleConfig(enabled: false, schedules: []);
-    }
   }
 
   @override